refactor: LottoGame 안에 남아있는 view 함수 호출 제거
